Move LSP summary modal to admin view
